// frontend/src/types/EOD.ts
export interface EOD {
  _id: any;
  register: string;
  sequence: string;
  opened: Date;
  closed: Date;
  xeroAmount: number;
  xeroPosted: string;
  xeroCreditNoteAmount: number;
  xeroCreditNotePosted: string;
  giftCardAmount: number;
  giftCardPosted: string;
  cashRoundingAmount: number;
  cashRoundingPosted: string;
  eftposCommbankAmount: number;
  eftposCommbankPosted: string;
  smartpayAmount: number;
  smartpayPosted: string;
  shopbackAmount: number;
  shopbackPosted: string;
  lightspeedPaymentsAmount: number;
  lightspeedPaymentsPosted: string;
  amexAmount: number;
  amexPosted: string;
  squareAmount: number;
  squarePosted: string;
  storeCreditAmount: number;
  storeCreditPosted: string;
  zipPayAmount: number;
  zipPayPosted: string;
  otherPaymentMethodAmount: number;
  otherPaymentMethodPosted: string;
  afterpayManualAmount: number;
  afterpayManualPosted: string;
  cashAmount: number;
  cashPosted: string;
  zellerT1Amount: number;
  zellerT1Posted: string;
  eftposNewAfterpayAmount: number;
  eftposNewAfterpayPosted: string;
  total: number;
}